requirements.txt
discord.py==2.3.2
asyncio==3.4.3
tinydb==4.8.0